+++ app/Api/Controllers/APIController.php
@@ -31,10 +31,10 @@ class ApiController extends Controller
     public function get_stats()
     {
         $stats               = array();
-        $stats['devices']    = \App\Device::all()->count();
-        $stats['ports']      = \App\Port::all()->count();
-        $stats['syslog']     = \App\Syslog::all()->count();
-        $stats['eventlog']   = \App\Eventlog::all()->count();
+        $stats['devices']    = \App\Models\Device::all()->count();
+        $stats['ports']      = \App\Models\Port::all()->count();
+        $stats['syslog']     = \App\Models\Syslog::all()->count();
+        $stats['eventlog']   = \App\Models\Eventlog::all()->count();
         $stats['apps']       = DB::table('applications')->count();;
         $stats['services']   = DB::table('services')->count();
         $stats['storage']    = DB::table('storage')->count();
