+++ tests/general/AboutPageTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\User;
+use App\Models\User;
 
 class AboutPageTest extends TestCase
 {
