+++ tests/general/NotificationPageTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\User;
+use App\Models\User;
 
 class NotificationPageTest extends TestCase
 {
